[Docs] Clean Code 개선
